fixed exercise1/compute_cube_index and exercise1/marching_cubes

+++ exercise1/exercise_1/occupancy_grid.py
@@ -12,10 +12,10 @@ def occupancy_grid(sdf_function, resolution):
     :return: An occupancy grid of specified resolution (i.e. an array of dim (resolution, resolution, resolution)
              with value 0 outside the shape and 1 inside.
     """
-    ####################################################################
-    x_ = np.linspace(start=-0.5, stop=0.5, num=resolution, endpoint=True, dtype=np.float64)
-    y_ = np.linspace(start=-0.5, stop=0.5, num=resolution, endpoint=True, dtype=np.float64)
-    z_ = np.linspace(start=-0.5, stop=0.5, num=resolution, endpoint=True, dtype=np.float64)
+    #####################################################################################
+    x_ = np.linspace(start=-0.5, stop=0.5, num=resolution, endpoint=True)
+    y_ = np.linspace(start=-0.5, stop=0.5, num=resolution, endpoint=True)
+    z_ = np.linspace(start=-0.5, stop=0.5, num=resolution, endpoint=True)
     xx, yy, zz = np.meshgrid(x_, y_, z_, indexing='ij')
     x, y, z = xx.flatten(), yy.flatten(), zz.flatten()
 
@@ -26,4 +26,4 @@ def occupancy_grid(sdf_function, resolution):
     occupancy_grid[sdf_grid < 0] = 1  # negative values are inside 
 
     return occupancy_grid
-    ####################################################################
+    #####################################################################################
